Some fixes and element improvements.
